Separate out base functionality of Storage class. Create separate localStorage and sessionStorage with different persistence behaviour across sessions.
